move config to separate files, add more examples, add `develop` build version
